requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.5.2; python_version >= '3.7'
certifi==2022.6.15; python_version >= '3.6'
cffi==1.15.1
charset-normalizer==2.1.0; python_version >= '3.6'
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4; python_version >= '3.6'
defusedxml==0.7.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
django-debug-toolbar==3.5.0
django-filter==22.1
django-templated-mail==1.1.1
django==4.0.6
djangorestframework-simplejwt==4.8.0
djangorestframework==3.13.1
djoser==2.1.0
idna==3.3; python_version >= '3.5'
itypes==1.2.0
jinja2==3.1.2; python_version >= '3.7'
markupsafe==2.1.1; python_version >= '3.7'
oauthlib==3.2.0; python_version >= '3.6'
pycparser==2.21
pyjwt==2.4.0; python_version >= '3.6'
python3-openid==3.2.0
pytz==2022.1
requests-oauthlib==1.3.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
requests==2.28.1; python_version >= '3.7' and python_full_version < '4.0.0'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
social-auth-app-django==4.0.0
social-auth-core==4.3.0; python_version >= '3.6'
sqlparse==0.4.2; python_version >= '3.5'
uritemplate==4.1.1; python_version >= '3.6'
urllib3==1.26.11; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5' and python_full_version < '4.0.0'
